requirements.txt
aiogram>=3.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.28.0
psycopg2-binary>=2.9.0
beautifulsoup4>=4.12.0
requests>=2.31.0
cryptography>=41.0.0
celery[redis]>=5.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
aiohttp>=3.9.0
kombu>=5.3.0
redis>=5.0.0
